test::proptest::number_to_integer_with_float
