+++ lumen_runtime/src/otp/erlang/ceil_1/test.rs
@@ -1,16 +1,6 @@
-use std::convert::TryInto;
-
-use num_bigint::BigInt;
-
-use num_traits::Num;
-
-use proptest::strategy::Just;
-use proptest::{prop_assert, prop_assert_eq};
-
-use liblumen_alloc::erts::term::prelude::{Encoded, Float};
+use proptest::prop_assert;
 
 use crate::otp::erlang::ceil_1::native;
-use crate::test::strategy;
 
 #[test]
 fn without_number_errors_badarg() {
@@ -24,37 +14,9 @@ fn with_integer_returns_integer() {
 
 #[test]
 fn with_float_round_up_to_next_integer() {
-    run!(
-        |arc_process| {
-            (
-                Just(arc_process.clone()),
-                strategy::term::float(arc_process.clone()),
-            )
-        },
-        |(arc_process, number)| {
-            let result = native(&arc_process, number);
+    crate::test::number_to_integer_with_float(file!(), native, |number, _, result_term| {
+        prop_assert!(number <= result_term, "{:?} <= {:?}", number, result_term);
 
-            prop_assert!(result.is_ok());
-
-            let result_term = result.unwrap();
-
-            prop_assert!(result_term.is_integer());
-
-            let number_float: Float = number.try_into().unwrap();
-            let number_f64: f64 = number_float.into();
-
-            if number_f64.fract() == 0.0 {
-                // f64::to_string() has no decimal point when there is no `fract`.
-                let number_big_int =
-                    <BigInt as Num>::from_str_radix(&number_f64.to_string(), 10).unwrap();
-                let result_big_int: BigInt = result_term.try_into().unwrap();
-
-                prop_assert_eq!(number_big_int, result_big_int);
-            } else {
-                prop_assert!(number <= result_term, "{:?} <= {:?}", number, result_term);
-            }
-
-            Ok(())
-        },
-    );
+        Ok(())
+    })
 }
