+++ lumen_runtime/src/test/proptest.rs
@@ -3,6 +3,10 @@ use std::convert::TryInto;
 use std::sync::atomic::AtomicUsize;
 use std::sync::Arc;
 
+use num_bigint::BigInt;
+
+use num_traits::Num;
+
 use proptest::strategy::{Just, Strategy};
 use proptest::test_runner::{Config, TestCaseResult, TestRunner};
 use proptest::{prop_assert, prop_assert_eq};
@@ -131,6 +135,47 @@ pub fn monitor_count(process: &Process) -> usize {
 
 pub fn monitored_count(process: &Process) -> usize {
     process.monitored_pid_by_reference.lock().len()
+}
+
+pub fn number_to_integer_with_float(
+    source_file: &'static str,
+    native: fn(&Process, Term) -> exception::Result<Term>,
+    non_zero_assertion: fn(Term, f64, Term) -> TestCaseResult,
+) {
+    run(
+        source_file,
+        |arc_process| {
+            (
+                Just(arc_process.clone()),
+                super::strategy::term::float(arc_process.clone()),
+            )
+        },
+        |(arc_process, number)| {
+            let result = native(&arc_process, number);
+
+            prop_assert!(result.is_ok());
+
+            let result_term = result.unwrap();
+
+            prop_assert!(result_term.is_integer());
+
+            let number_float: Float = number.try_into().unwrap();
+            let number_f64: f64 = number_float.into();
+
+            if number_f64.fract() == 0.0 {
+                // f64::to_string() has no decimal point when there is no `fract`.
+                let number_big_int =
+                    <BigInt as Num>::from_str_radix(&number_f64.to_string(), 10).unwrap();
+                let result_big_int: BigInt = result_term.try_into().unwrap();
+
+                prop_assert_eq!(number_big_int, result_big_int);
+
+                Ok(())
+            } else {
+                non_zero_assertion(number, number_f64, result_term)
+            }
+        },
+    );
 }
 
 pub fn process(parent_process: &Process, options: Options) -> Spawned {
